회원 및 프로필 정보 수정 #73

+++ src/main/java/com/project/date/dto/request/ProfileRequestDto.java
@@ -8,8 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 public class ProfileRequestDto {
-
-    private Long age;
+    private int age;
     private String mbti;
     private String introduction;
     private String idealType;
